deepfloat serialisation; tabulated activation serialisation

+++ util.go
@@ -24,7 +24,7 @@ func Sum(xx []Deepfloat64) (sum Deepfloat64) {
 
 // Round to nearest integer
 func Round(x Deepfloat64) Deepfloat64 {
-	return DF(math.Floor(x + .5))
+	return DF(math.Floor(float64(x) + .5))
 }
 
 // fibonacci returns a function that returns
